core: implement public line update routines
Line info is retrieved from the kernel for the first time
automatically when gpiod_chip_get_line() is called. Then the line
is updated automatically on request/release. If the ioctl fails in the
latter case, the up_to_date flag is set to false.

Expose an interface allowing users to check if a line is up-to-date
and manually update it if needed.

Signed-off-by: Bartosz Golaszewski <bartekgola@gmail.com>