tests: don't hardcode the chip names
Use test_chip_name() to get the name of a chip. Don't hardcode it as
it will make the tests not work if any gpiochips other than the ones
created by gpio-mockup already exist in the system.

This also fixes a couple hidden bugs, where we requested inexistent
chips in tools' command-line arguments.

Signed-off-by: Bartosz Golaszewski <bartekgola@gmail.com>