core: separate the chip and line logic
Lazily allocate memory for line objects in chip.c. This eliminates the
need for chip routines to know the size of a line object.

Line objects now contain the chip file descriptor as well, so that we
don't need to access the chip object in any way from line.c.

Signed-off-by: Bartosz Golaszewski <bartekgola@gmail.com>