tests: re-enable a test case for gpiomon
The test case in which we verify if falling-edge events are correctly
ignored was failing due to a bug in the kernel. This has now been
fixed with commit df1e76f28ffe ("gpiolib: skip unwanted events, don't
convert them to opposite edge") and released in linux v4.12.6.

Re-enable the test and make this version a hard requirement for
running the test suite.

Signed-off-by: Bartosz Golaszewski <bartekgola@gmail.com>