bindings: python: bail-out if the line was not found in gpiofind.py
Currently we're not checking if the object returned from
gpiod.fine_line() is None which results in an attribute error being
raised when we try to call line methods on a non-line object.

Add an if and bail-out with an error code if the line with given name
was not found.

Signed-off-by: Bartosz Golaszewski <bartekgola@gmail.com>