bindings: python: add support for 'default_val' in Line.request()
Until now Line.request() only took 'default_vals' argument which was
passed down directly to LineBulk.request() internally. This meant that
even when requesting a single line in output mode, the default value
had to be passed in a sequence.

Add support for the 'default_val' argument which makes it possible to
specify the default value as an integer. The previous 'default_vals'
argument is still supported, but the two cannot be passed at the same
time.

Documentation for default_vals is removed, as it's now deprecated for
Line.request().

Signed-off-by: Bartosz Golaszewski <bgolaszewski@baylibre.com>