tests: setup libgpiomockup before checking the kernel version
The gpio-mockup helper library checks the kernel version itself when
creating the context object. We then check the version again (as the
minimum versions for libgpiomockup and the test-suite may differ) in
the test-suite's main function. Make sure the library runs the check
first as we don't want to tell the user the version is OK and then
fail because it isn't...

Signed-off-by: Bartosz Golaszewski <bgolaszewski@baylibre.com>