tools: tests: run the bats script indirectly from a shell script
Bats makes it very difficult to run some initial code once, perform
checks for requirements and optionally exit if they are not satisfied.

Instead of working around it in the .bats script, just create
a separate shell script that does all the initial checks and then
execs the actual bats test-suite.

Signed-off-by: Bartosz Golaszewski <bgolaszewski@baylibre.com>