tools: tests: fix the gpiomon test case with --active-low switch
Kernel commit 223ecaf140b1 ("gpiolib: fix incorrect IRQ requesting of
an active-low lineevent") fixed an issue with line events not being
inverted with the GPIOHANDLE_REQUEST_ACTIVE_LOW flag. This change has
now been released in linux v5.2.7 and the relevant test case for
gpiomon needs to be adjusted.

Since the line going down will now be interpreted as a rising edge
event, set the pull to 1 before running gpiomon and then set it to 0
after starting it to trigger a rising edge event.

The minimum kernel version to run tests must be set to v5.2.7.

Signed-off-by: Bartosz Golaszewski <bgolaszewski@baylibre.com>