tests: event: fix the active-low switch test case
Kernel commit 223ecaf140b1 ("gpiolib: fix incorrect IRQ requesting of
an active-low lineevent") fixed an issue with line events not being
inverted with the GPIOHANDLE_REQUEST_ACTIVE_LOW flag. This change has
now been released in linux v5.2.7 and the relevant test case must be
adjusted.

Convert the test case to expect a falling edge event. Since the first
pull change will be 0->1, it will be interpreted as such when the
active-low flag is used.

The minimum kernel version to run the core test-suite is now set to
v5.2.7.

Signed-off-by: Bartosz Golaszewski <bgolaszewski@baylibre.com>