tools: tests: restore the bats shebang in the bats source file
In commit f52e12eafd17 ("tools: tests: run the bats script indirectly
from a shell script") we removed the bats shebang and dropped the
execute permission from the .bats test-suite file.

While the preferred method of running the gpio-tools tests is using
the helper bash script, the .bats file should still be executable on
its own. Restore the execute permissions and the shebang.

Signed-off-by: Bartosz Golaszewski <bgolaszewski@baylibre.com>