core: rename up_to_date to needs_update in struct gpiod_line
This variable is called up_to_date which can lead readers to the wrong
conclusion that libgpiod can somehow know when external actors (such as
a different process or a kernel driver) change the state of a GPIO line.

This is not the case and libgpiod only ever knows if a line needs info
update when a previous internal call to gpiod_line_update() fails.

The kernel does not provide any notification mechanism currently.

For that reason: rename the variable to needs_update.

While at it: clarify the documentation for gpiod_line_needs_update().

Signed-off-by: Bartosz Golaszewski <bgolaszewski@baylibre.com>