build: check linux headers version at build-time
Original libgpiod relied on linux headers v4.8 to build but it was only
documented in README and not enforced at build-time. We now support
features first available in linux v5.5. Add a check to configure.ac that
verifies if recent enough kernel headers are available - otherwise we'd
fail only when trying to build the library.

Signed-off-by: Bartosz Golaszewski <bgolaszewski@baylibre.com>