tests: line: fix line request flags
In the test case meant to make sure we can't make a request with a bulk
object holding lines from different chips, we're using an invalid value
in the flags field of the request structure. This alone may cause the
request call to fail with EINVAL so this test case is buggy. Use the
correct constant for the active-low flag.

Fixes: b7c5026e51e4 ("tests: add a test case for a bad line request")
Signed-off-by: Bartosz Golaszewski <bgolaszewski@baylibre.com>