bindings: cxx: remove duplicate mapping helper
We have two helpers that do the same thing: given a key and a map, they
return the associated value or throw an exception if the map doesn't
contain it. The difference is just in the type of the exception. Remove
one and reuse the other. While at it: modify the names a bit for better
readability - especially the mapping function which is not limited to
just mapping enum types.

Signed-off-by: Bartosz Golaszewski <bartosz.golaszewski@linaro.org>